requirements.txt
anthropic>=0.42.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
pydantic-settings>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
click>=8.1.0
python-dotenv>=1.0.0
